dbus server: move the listen method to TuhiDBusServer
The listen() method should not attempt to connect to a particular device,
but instead it should open the daemon to any solicitation from the device.

Once a device solicits the deamon (i.e. the physical sync button has been
pressed by the user), the deamon should connect to the device, retrieve
the data and disconnects itself. This will come in a later commit.

Signed-off-by: Benjamin Tissoires <benjamin.tissoires@gmail.com>