do not attempt to connect to a device if we are not sure we can
When the device appears from the daemon point of view, there is no
reason that it is actually able to handle incoming connections.
We should solely rely on the discovery mode, so not try to attempt
to connect on device-added.

Also, if the RSSI gets invalidated, it just means the discovery mode
stopped, not that the device is ready.

Signed-off-by: Benjamin Tissoires <benjamin.tissoires@gmail.com>